Group account-related requests
