+++ app/Http/Controllers/UserController.php
@@ -12,9 +12,9 @@ use App\Accounts\UserRole;
 use App\Accounts\UserRepository;
 use App\Accounts\UserUpdater;
 use App\Accounts\UserUpdaterListener;
-use App\Http\Requests\UpdateProfileRequest;
-use App\Http\Requests\UpdateUserEmailRequest;
-use App\Http\Requests\UpdateUserPasswordRequest;
+use App\Http\Requests\Accounts\UpdateProfileRequest;
+use App\Http\Requests\Accounts\UpdateUserEmailRequest;
+use App\Http\Requests\Accounts\UpdateUserPasswordRequest;
 use Hash;
 use Input;
 use Redirect;
