+++ app/Http/Requests/Accounts/UpdateUserEmailRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Accounts;
 
 use App\Http\Requests\Request;
 
-class UpdateProfileRequest extends Request
+class UpdateUserEmailRequest extends Request
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,8 +24,7 @@ class UpdateProfileRequest extends Request
     public function rules()
     {
         return [
-            'first_name' => 'required',
-            'last_name' => 'required'
+            'email' => 'required|email|unique:users,email,' . $this->user()->id
         ];
     }
 
@@ -37,8 +36,7 @@ class UpdateProfileRequest extends Request
     public function messages()
     {
         return [
-            'firstname.required' => 'The first name field is required.',
-            'lastname.required' => 'The last name field is required.'
+            'email.unique' => 'This email is already used'
         ];
     }
 }
