+++ app/Http/Requests/Accounts/UpdateUserPasswordRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Accounts;
 
 use App\Http\Requests\Request;
